DonutChart: Migrated tests from Enzyme to react tests
